Mini Rev Bot
This commit includes the Mini Rev module robot.  it's a 200g plastic
chassis for use with the new Rev module.  It can be used with two tetrix
motors, as well as two color sensors for following lines.  It also has
tetrix holes for a phone mount, as well as a front facing motor.